+++ scripts/calibnet_health_check.sh
@@ -71,11 +71,14 @@ du -hS ~/.local/share/forest/calibnet
 $FOREST_CLI_PATH db gc
 du -hS ~/.local/share/forest/calibnet
 
-echo "Exporting snapshot"
+echo "Exporting uncompressed snapshot"
 $FOREST_CLI_PATH snapshot export
 
 echo "Verifing snapshot checksum"
 sha256sum -c ./*.sha256sum
+
+echo "Exporting zstd compressed snapshot"
+$FOREST_CLI_PATH snapshot export --compressed
 
 echo "Testing js console"
 $FOREST_CLI_PATH attach --exec 'showPeers()'
@@ -89,8 +92,11 @@ $FOREST_CLI_PATH --chain mainnet snapshot validate "$SNAPSHOT_DIRECTORY"/*.car -
 }
 set -e
 
-echo "Validating as calibnet snapshot"
+echo "Validating as calibnet snapshot (uncompressed)"
 $FOREST_CLI_PATH --chain calibnet snapshot validate "$SNAPSHOT_DIRECTORY"/*.car --force
+
+echo "Validating as calibnet snapshot (compressed)"
+$FOREST_CLI_PATH --chain calibnet snapshot validate "$SNAPSHOT_DIRECTORY"/*.zst --force
 
 echo "Print forest log files"
 ls -hl "$LOG_DIRECTORY"