+++ CHANGELOG.md
@@ -63,6 +63,9 @@
   zstd compressed snapshot car files
 - [#2741](https://github.com/ChainSafe/forest/issues/2741): Support fetching
   zstd compressed snapshots with filecoin provider via `--compressed` option
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support exporting
+  zstd compressed snapshots via `--compressed` option in
+  `forest-cli snapshot export` subcommand
 - [#1454](https://github.com/ChainSafe/forest/issues/1454): Added state
   migration support for NV18.